+++ tests/integration/cli_build_test.sh
@@ -123,13 +123,14 @@ set -- build tool --check \
   --build-architecture x86_64 \
   --target-architecture x86_64 \
   --start "$nonce" \
+  --build-parallelism 1 \
+  --build-source-date-epoch 0 \
   --runtime-root "$runtime" \
   --build-root "$build" \
   --artifact-root "$runtime/content" \
   --interpreter "$interpreter" \
   --build-user-id "$uid" \
   --build-group-id "$gid" \
-  --source-date-epoch 0 \
   --max-steps 1
 for group in $groups; do
   if [ "$group" != "$gid" ]; then
@@ -156,13 +157,14 @@ set -- build tool --check \
   --build-architecture x86_64 \
   --target-architecture x86_64 \
   --start "$nonce" \
+  --build-parallelism 1 \
+  --build-source-date-epoch 0 \
   --runtime-root "$runtime" \
   --build-root "$build" \
   --artifact-root "$artifacts" \
   --interpreter "$interpreter" \
   --build-user-id "$uid" \
   --build-group-id "$gid" \
-  --source-date-epoch 0 \
   --max-steps 1
 for group in $groups; do
   if [ "$group" != "$gid" ]; then
@@ -213,7 +215,6 @@ set -- build \
   --interpreter "$interpreter" \
   --build-user-id "$uid" \
   --build-group-id "$gid" \
-  --source-date-epoch 0 \
   --max-steps 2
 for group in $groups; do
   if [ "$group" != "$gid" ]; then
@@ -241,7 +242,6 @@ set -- build \
   --interpreter "$interpreter" \
   --build-user-id "$uid" \
   --build-group-id "$gid" \
-  --source-date-epoch 0 \
   --max-steps 2
 for group in $groups; do
   if [ "$group" != "$gid" ]; then
@@ -357,7 +357,6 @@ set -- build \
   --interpreter "$interpreter" \
   --build-user-id "$uid" \
   --build-group-id "$gid" \
-  --source-date-epoch 0 \
   --max-steps 3
 for group in $groups; do
   if [ "$group" != "$gid" ]; then