+++ tests/integration/cli_run_test.sh
@@ -70,6 +70,9 @@ run_command()
       --build-architecture x86_64 \
       --target-architecture x86_64 \
       --goal 'run=@base'
+    set -- "$@" \
+      --build-parallelism 1 \
+      --build-source-date-epoch 0
   elif [ "$inject_resume_semantics" = yes ]; then
     set -- "$@" --goal 'run=@base'
   fi
@@ -84,7 +87,6 @@ run_command()
     --build-group-id "$build_gid" \
     --lifecycle-user-id "$uid" \
     --lifecycle-group-id "$gid" \
-    --source-date-epoch 0 \
     --max-steps "$maximum_steps"
   for group in $build_groups; do
     if [ "$group" != "$build_gid" ]; then